Version 0.7.2
Fixed 2 bugs in script 983 related to file_transport_primary_team_displayname and file_transport_primary_team_user_membership when set to true - the display name and team members where not updated in the target when other file parameters where false this is now fixed.
Fixed 2 bugs related to error handling of updating the target team. This is now fixed.