Update Source to Target - Em - Scripts 9xx.postman_environment.json
With NEO the accesstoken is required to contain a value, even an invalid value for the correct response code to be returned by the SAC API. This update simply adds a dummy value into the accesstoken value preventing a problem on the NEO platform when the Test scripts are not run beforehand. It also solves the problem when the environment is exported to be used by newman and the environment values are not persisted.