marketingemailoptin removed
For the users' endpoint: marketingemailoptin is a legacy attribute that is no longer used. If it is present in the request body in the future, it will cause that request to fail. Thus, it is being removed now to prevent a failed request in the future. Given that it has no purpose today, it's very safe to remove it